Create CelestronControlApp.py
Initial commit. It needs better event handling since the Clock.schedule_interval slows things way down and makes the buttons slow to react. The multibyte output needs to be fixed. SYNC TIME and Get Time buttons do not work yet.  The nexstar.py Nexstar Class also needs attention.  This is intended to run on a Raspberry Pi attached to to a NexStar SE Telescope Mount as a remote control.